feat(context): consolidate context operations into appium_context
Merge appium_get_contexts and appium_switch_context into a single action-based tool with list/switch actions, and update UI/docs references to the unified command.
