--- src/tools/context/switch-context.ts
@@ -1,110 +1,0 @@
-import { FastMCP } from 'fastmcp';
-import { z } from 'zod';
-import {
-  getDriver,
-  isRemoteDriverSession,
-  setCurrentContext,
-} from '../../session-store.js';
-import { getContexts, getCurrentContext, setContext } from '../../command.js';
-
-export default function switchContext(server: FastMCP): void {
-  const schema = z.object({
-    context: z
-      .string()
-      .describe(
-        'The name of the context to switch to. Common values: "NATIVE_APP" for native context, or "WEBVIEW_<id>" / "WEBVIEW_<package>" for webview contexts.'
-      ),
-    sessionId: z
-      .string()
-      .optional()
-      .describe('Session ID to target. If omitted, uses the active session.'),
-  });
-
-  server.addTool({
-    name: 'appium_switch_context',
-    description:
-      'Switch to a specific context in the Appium session. Use this to switch between native app context (NATIVE_APP) and webview contexts (WEBVIEW_<id> or WEBVIEW_<package>). Use appium_get_contexts to see available contexts first.',
-    parameters: schema,
-    annotations: {
-      readOnlyHint: false,
-      openWorldHint: false,
-    },
-    execute: async (args: any, _context: any): Promise<any> => {
-      const driver = getDriver(args.sessionId);
-      if (!driver) {
-        throw new Error('No driver found. Please create a session first.');
-      }
-
-      if (isRemoteDriverSession(driver)) {
-        throw new Error(
-          'Get context is not yet implemented for the remote driver'
-        );
-      }
-
-      try {
-        const [currentContext, availableContexts] = await Promise.all([
-          getCurrentContext(driver).catch(() => null),
-          getContexts(driver).catch(() => [] as string[]),
-        ]);
-
-        if (currentContext === args.context) {
-          return {
-            content: [
-              {
-                type: 'text',
-                text: `Already on context "${args.context}".`,
-              },
-            ],
-          };
-        }
-
-        if (!availableContexts || availableContexts.length === 0) {
-          return {
-            content: [
-              {
-                type: 'text',
-                text: 'No contexts available. Cannot switch context.',
-              },
-            ],
-            isError: true,
-          };
-        }
-
-        if (!availableContexts.includes(args.context)) {
-          return {
-            content: [
-              {
-                type: 'text',
-                text: `Context "${args.context}" not found. Available contexts: ${JSON.stringify(availableContexts, null, 2)}`,
-              },
-            ],
-            isError: true,
-          };
-        }
-        await setContext(driver, args.context);
-        // Verify the switch was successful
-        const newContext = await getCurrentContext(driver);
-        setCurrentContext(newContext);
-
-        return {
-          content: [
-            {
-              type: 'text',
-              text: `Successfully switched context from "${currentContext}" to "${newContext}".`,
-            },
-          ],
-        };
-      } catch (err: any) {
-        return {
-          content: [
-            {
-              type: 'text',
-              text: `Failed to switch context. Error: ${err.toString()}`,
-            },
-          ],
-          isError: true,
-        };
-      }
-    },
-  });
-}
