+++ src/tools/index.ts
@@ -70,8 +70,7 @@ import isAppInstalled from './app-management/is-app-installed.js';
 import queryAppState from './app-management/query-app-state.js';
 import clearApp from './app-management/clear-app.js';
 import deepLink from './app-management/deep-link.js';
-import getContexts from './context/get-contexts.js';
-import switchContext from './context/switch-context.js';
+import context from './context/context.js';
 
 export default function registerTools(server: FastMCP): void {
   // Wrap addTool to inject logging around tool execution
@@ -206,8 +205,7 @@ export default function registerTools(server: FastMCP): void {
   deepLink(server);
 
   // Context Management
-  getContexts(server);
-  switchContext(server);
+  context(server);
 
   // Test Generation
   generateLocators(server);
